refactor: extract NativeBinary from NativeLibrary
